Added option to use URL query string as path
Optional. I think this is a nice option, since this makes things more flexible. For instance you could place the handler script anywhere and have paths relative to it instead of root. Though, it is not possible to use the query string as a source of parameters UNLESS YOU USE A REWRITE. You should also change the RewriteBase to match the location of the script.

Does not change any other behavior and is 100% safe to use with existing scripts.